refacto

+++ pharmaSmart-app/src/main/webapp/app/entities/customer/customer.component.ts
@@ -1,49 +1,50 @@
-import { Component, inject, OnDestroy, OnInit, viewChild } from '@angular/core';
-import { HttpHeaders, HttpResponse } from '@angular/common/http';
-import { ActivatedRoute, Data, ParamMap, Router, RouterLink } from '@angular/router';
-import { combineLatest, Observable, Subject } from 'rxjs';
-import { takeUntil } from 'rxjs/operators';
-import { NgbModal } from '@ng-bootstrap/ng-bootstrap';
-import { ICustomer } from 'app/shared/model/customer.model';
-import { ITEMS_PER_PAGE } from 'app/shared/constants/pagination.constants';
-import { CustomerService } from './customer.service';
-import { MenuItem, MessageService } from 'primeng/api';
-import { UninsuredCustomerFormComponent } from './uninsured-customer-form/uninsured-customer-form.component';
-import { TranslateService } from '@ngx-translate/core';
-import { FormAyantDroitComponent } from './form-ayant-droit/form-ayant-droit.component';
-import { WarehouseCommonModule } from '../../shared/warehouse-common/warehouse-common.module';
-import { FormsModule } from '@angular/forms';
-import { ButtonModule } from 'primeng/button';
-import { RippleModule } from 'primeng/ripple';
-import { InputTextModule } from 'primeng/inputtext';
-import { TableLazyLoadEvent, TableModule } from 'primeng/table';
-import { DialogModule } from 'primeng/dialog';
-import { FileUploadModule } from 'primeng/fileupload';
-import { ToolbarModule } from 'primeng/toolbar';
-import { SplitButtonModule } from 'primeng/splitbutton';
-import { TooltipModule } from 'primeng/tooltip';
-import { CustomerTiersPayantComponent } from './customer-tiers-payant/customer-tiers-payant.component';
-import { IClientTiersPayant } from '../../shared/model/client-tiers-payant.model';
-import { AssureFormStepComponent } from './assure-form-step/assure-form-step.component';
-import { PrimeNG } from 'primeng/config';
-import { Select, SelectModule } from 'primeng/select';
-import { IconField } from 'primeng/iconfield';
-import { InputIcon } from 'primeng/inputicon';
-import { DividerModule } from 'primeng/divider';
-import { CustomerCarnetComponent } from './carnet/customer-carnet.component';
-import { ConfirmDialogComponent } from '../../shared/dialog/confirm-dialog/confirm-dialog.component';
-import { showCommonModal } from '../sales/selling-home/sale-helper';
-import { SpinnerComponent } from '../../shared/spinner/spinner.component';
-import { Card } from 'primeng/card';
-import { FloatLabel } from 'primeng/floatlabel';
+import { Component, inject, OnDestroy, OnInit, viewChild } from "@angular/core";
+import { HttpHeaders, HttpResponse } from "@angular/common/http";
+import { ActivatedRoute, Data, ParamMap, Router, RouterLink } from "@angular/router";
+import { combineLatest, Observable, Subject } from "rxjs";
+import { takeUntil } from "rxjs/operators";
+import { NgbModal } from "@ng-bootstrap/ng-bootstrap";
+import { ICustomer } from "app/shared/model/customer.model";
+import { ITEMS_PER_PAGE } from "app/shared/constants/pagination.constants";
+import { CustomerService } from "./customer.service";
+import { MenuItem } from "primeng/api";
+import { UninsuredCustomerFormComponent } from "./uninsured-customer-form/uninsured-customer-form.component";
+import { TranslateService } from "@ngx-translate/core";
+import { FormAyantDroitComponent } from "./form-ayant-droit/form-ayant-droit.component";
+import { FormsModule } from "@angular/forms";
+import { ButtonModule } from "primeng/button";
+import { RippleModule } from "primeng/ripple";
+import { InputTextModule } from "primeng/inputtext";
+import { TableLazyLoadEvent, TableModule } from "primeng/table";
+import { DialogModule } from "primeng/dialog";
+import { FileUploadModule } from "primeng/fileupload";
+import { ToolbarModule } from "primeng/toolbar";
+import { SplitButtonModule } from "primeng/splitbutton";
+import { TooltipModule } from "primeng/tooltip";
+import { CustomerTiersPayantComponent } from "./customer-tiers-payant/customer-tiers-payant.component";
+import { IClientTiersPayant } from "../../shared/model";
+import { AssureFormStepComponent } from "./assure-form-step/assure-form-step.component";
+import { PrimeNG } from "primeng/config";
+import { Select, SelectModule } from "primeng/select";
+import { IconField } from "primeng/iconfield";
+import { InputIcon } from "primeng/inputicon";
+import { DividerModule } from "primeng/divider";
+import { CustomerCarnetComponent } from "./carnet/customer-carnet.component";
+import { showCommonModal } from "../sales/selling-home/sale-helper";
+import { SpinnerComponent } from "../../shared/spinner/spinner.component";
+import { Card } from "primeng/card";
+import { FloatLabel } from "primeng/floatlabel";
+import { Toast } from "primeng/toast";
+import { CommonModule } from "@angular/common";
+import { NgbConfirmDialogService } from "../../shared/dialog/ngb-confirm-dialog/ngb-confirm-dialog.directive";
+import { NotificationService } from "../../shared/services/notification.service";
 
 @Component({
-  selector: 'jhi-customer',
-  templateUrl: './customer.component.html',
-  styleUrl: './customer.component.scss',
-  providers: [MessageService],
+  selector: "app-customer",
+  templateUrl: "./customer.component.html",
+  styleUrls: ["./customer.component.scss"],
   imports: [
-    WarehouseCommonModule,
+    CommonModule,
     SelectModule,
     ToolbarModule,
     FormsModule,
@@ -60,24 +61,24 @@ import { FloatLabel } from 'primeng/floatlabel';
     Select,
     IconField,
     InputIcon,
-    ConfirmDialogComponent,
     SpinnerComponent,
     Card,
     FloatLabel,
-  ],
+    Toast
+  ]
 })
 export class CustomerComponent implements OnInit, OnDestroy {
   translate = inject(TranslateService);
   primeNGConfig = inject(PrimeNG);
   customers?: ICustomer[];
-  types: string[] = ['TOUT', 'ASSURE', 'STANDARD'];
+  types: string[] = ["TOUT", "ASSURE", "STANDARD"];
   statuts: object[] = [
-    { value: 'ENABLE', label: 'Actifs' },
-    { value: 'DISABLE', label: 'Désactivés' },
+    { value: "ENABLE", label: "Actifs" },
+    { value: "DISABLE", label: "Désactivés" }
   ];
-  typeSelected = '';
-  statutSelected = 'ENABLE';
-  search = '';
+  typeSelected = "";
+  statutSelected = "ENABLE";
+  search = "";
   totalItems = 0;
   itemsPerPage = ITEMS_PER_PAGE;
   page!: number;
@@ -94,37 +95,33 @@ export class CustomerComponent implements OnInit, OnDestroy {
   protected router = inject(Router);
   private readonly modalService = inject(NgbModal);
   private destroy$ = new Subject<void>();
-  private readonly spinner = viewChild.required<SpinnerComponent>('spinner');
-  private readonly messageService = inject(MessageService);
-  private readonly confimDialog = viewChild.required<ConfirmDialogComponent>('confirmDialog');
+  private readonly spinner = viewChild.required<SpinnerComponent>("spinner");
+  private readonly confirmDialog = inject(NgbConfirmDialogService);
+  private readonly notificationService = inject(NotificationService);
+
 
   constructor() {
-    this.translate.use('fr');
-    this.translate.stream('primeng').subscribe(data => {
-      this.primeNGConfig.setTranslation(data);
-    });
-
     this.newCustomerbuttons = [
       {
-        label: 'Assuré',
-        icon: 'pi pi-user-plus',
-        command: () => this.addAssureCustomer('ASSURANCE'),
+        label: "Assuré",
+        icon: "pi pi-user-plus",
+        command: () => this.addAssureCustomer("ASSURANCE")
       },
       {
-        label: 'Carnet',
-        icon: 'pi pi-user-plus',
-        command: () => this.addCarnet('CARNET'),
+        label: "Carnet",
+        icon: "pi pi-user-plus",
+        command: () => this.addCarnet("CARNET")
       },
       {
-        label: 'Dépôt',
-        icon: 'pi pi-user-plus',
-        command: () => this.addCarnet('DEPOT'),
+        label: "Dépôt",
+        icon: "pi pi-user-plus",
+        command: () => this.addCarnet("DEPOT")
       },
       {
-        label: 'Standard',
-        icon: 'pi pi-user-plus',
-        command: () => this.addUninsuredCustomer(),
-      },
+        label: "Standard",
+        icon: "pi pi-user-plus",
+        command: () => this.addUninsuredCustomer()
+      }
     ];
   }
 
@@ -150,12 +147,12 @@ export class CustomerComponent implements OnInit, OnDestroy {
         sort: this.sort(),
         type: this.typeSelected,
         search: this.search,
-        status: this.statutSelected,
+        status: this.statutSelected
       })
       .pipe(takeUntil(this.destroy$))
       .subscribe({
         next: (res: HttpResponse<ICustomer[]>) => this.onSuccess(res.body, res.headers, pageToLoad, dontNavigate),
-        error: () => this.onError(),
+        error: () => this.onError()
       });
   }
 
@@ -170,19 +167,19 @@ export class CustomerComponent implements OnInit, OnDestroy {
           sort: this.sort(),
           type: this.typeSelected,
           search: this.search,
-          status: this.statutSelected,
+          status: this.statutSelected
         })
         .pipe(takeUntil(this.destroy$))
         .subscribe({
           next: (res: HttpResponse<ICustomer[]>) => this.onSuccess(res.body, res.headers, this.page, false),
-          error: () => this.onError(),
+          error: () => this.onError()
         });
     }
   }
 
   ngOnInit(): void {
-    this.typeSelected = 'TOUT';
-    this.statutSelected = 'ENABLE';
+    this.typeSelected = "TOUT";
+    this.statutSelected = "ENABLE";
     this.handleNavigation();
     this.loadPage();
     //   this.registerChangeInCustomers();
@@ -206,27 +203,27 @@ export class CustomerComponent implements OnInit, OnDestroy {
   }
 
   confirmRemove(customer: ICustomer): void {
-    this.confimDialog().onConfirm(
+    this.confirmDialog.onConfirm(
       () => {
-        if (customer.categorie === 'ASSURE') {
+        if (customer.categorie === "ASSURE") {
           this.deleteAssuredCustomer(customer);
         } else {
           this.delete(customer);
         }
       },
-      'SUPPRESSION DE CLIENT',
-      'Voulez-vous vraiment supprimer ce client ?',
+      "SUPPRESSION DE CLIENT",
+      "Voulez-vous vraiment supprimer ce client ?"
     );
   }
 
   confirmDesactivation(customer: ICustomer): void {
-    this.confimDialog().onConfirm(() => this.lock(customer), 'DESACTIVATION DE CLIENT', 'Voulez-vous vraiment désativer ce client ?');
+    this.confirmDialog.onConfirm(() => this.lock(customer), "DESACTIVATION DE CLIENT", "Voulez-vous vraiment désativer ce client ?");
   }
 
   sort(): string[] {
-    const result = [this.predicate + ',' + (this.ascending ? 'asc' : 'desc')];
-    if (this.predicate !== 'id') {
-      result.push('id');
+    const result = [this.predicate + "," + (this.ascending ? "asc" : "desc")];
+    if (this.predicate !== "id") {
+      result.push("id");
     }
     return result;
   }
@@ -238,14 +235,14 @@ export class CustomerComponent implements OnInit, OnDestroy {
       {
         entity: null,
         categorie,
-        title: `FORMULAIRE DE CREATION DE CLIENT [ ${categorie} ]`,
+        title: `FORMULAIRE DE CREATION DE CLIENT [ ${categorie} ]`
       },
       (resp: ICustomer) => {
         if (resp) {
           this.loadPage();
         }
       },
-      'xl',
+      "xl"
     );
   }
 
@@ -256,15 +253,15 @@ export class CustomerComponent implements OnInit, OnDestroy {
       {
         entity: null,
         typeAssure,
-        header: 'FORMULAIRE DE CREATION DE CLIENT ',
+        header: "FORMULAIRE DE CREATION DE CLIENT "
       },
       (resp: ICustomer) => {
         if (resp) {
           this.loadPage();
         }
       },
-      'xl',
-      'modal-dialog-80',
+      "xl",
+      "modal-dialog-80"
     );
   }
 
@@ -274,25 +271,25 @@ export class CustomerComponent implements OnInit, OnDestroy {
       AssureFormStepComponent,
       {
         entity: customer,
-        header: `FORMULAIRE DE MODIFICATION DE CLIENT  [ ${customer.fullName}  ]`,
+        header: `FORMULAIRE DE MODIFICATION DE CLIENT  [ ${customer.fullName}  ]`
       },
       (resp: ICustomer) => {
         if (resp) {
           this.loadPage();
         }
       },
-      'xl',
-      'modal-dialog-80',
+      "xl",
+      "modal-dialog-80"
     );
   }
 
   onRemoveTiersPayant(clientTiersPayant: IClientTiersPayant): void {
-    this.confimDialog().onConfirm(
+    this.confirmDialog.onConfirm(
       () => {
         this.handleServiceCall(this.customerService.deleteTiersPayant(clientTiersPayant.id), () => this.loadPage());
       },
-      'SUPPRESSION DE TIERS PAYANT',
-      'Voulez-vous vraiment supprimer ce tiers payant ?',
+      "SUPPRESSION DE TIERS PAYANT",
+      "Voulez-vous vraiment supprimer ce tiers payant ?"
     );
   }
 
@@ -302,14 +299,14 @@ export class CustomerComponent implements OnInit, OnDestroy {
       UninsuredCustomerFormComponent,
       {
         entity: null,
-        title: 'FORMULAIRE DE CREATION DE CLIENT ',
+        title: "FORMULAIRE DE CREATION DE CLIENT "
       },
       (resp: ICustomer) => {
         if (resp) {
           this.loadPage();
         }
       },
-      'xl',
+      "xl"
     );
   }
 
@@ -319,14 +316,14 @@ export class CustomerComponent implements OnInit, OnDestroy {
       UninsuredCustomerFormComponent,
       {
         entity: customer,
-        title: `FORMULAIRE DE MODIFICATION DE CLIENT  [ ${customer.fullName}  ]`,
+        title: `FORMULAIRE DE MODIFICATION DE CLIENT  [ ${customer.fullName}  ]`
       },
       (resp: ICustomer) => {
         if (resp) {
           this.loadPage();
         }
       },
-      'xl',
+      "xl"
     );
   }
 
@@ -337,14 +334,14 @@ export class CustomerComponent implements OnInit, OnDestroy {
       {
         entity: null,
         assure: customer,
-        title: "FORMULAIRE D'AJOUT D'AYANT DROIT ",
+        title: "FORMULAIRE D'AJOUT D'AYANT DROIT "
       },
       (resp: ICustomer) => {
         if (resp) {
           this.loadPage();
         }
       },
-      'xl',
+      "xl"
     );
   }
 
@@ -355,28 +352,28 @@ export class CustomerComponent implements OnInit, OnDestroy {
       {
         entity: ayantDroit,
         assure: customer,
-        title: `FORMULAIRE DE MODIFICATION D'AYANT DROIT [ ${ayantDroit.fullName}  ]`,
+        title: `FORMULAIRE DE MODIFICATION D'AYANT DROIT [ ${ayantDroit.fullName}  ]`
       },
       (resp: ICustomer) => {
         if (resp) {
           this.loadPage();
         }
-      },
+      }
     );
   }
 
   protected confirmRemoveAyantDroit(ayantDroit: ICustomer): void {
-    this.confimDialog().onConfirm(
+    this.confirmDialog.onConfirm(
       () => this.deleteAssuredCustomer(ayantDroit),
-      'SUPPRESSION',
-      'Voulez-vous vraiment supprimer cet ayant droit ?',
+      "SUPPRESSION",
+      "Voulez-vous vraiment supprimer cet ayant droit ?"
     );
   }
 
   protected onUploadJson(event: any): void {
     const formData: FormData = new FormData();
     const file = event.files[0];
-    formData.append('importjson', file, file.name);
+    formData.append("importjson", file, file.name);
     this.jsonDialog = false;
     this.handleServiceCall(this.customerService.uploadJsonData(formData), () => this.onPocesJsonSuccess());
   }
@@ -388,14 +385,14 @@ export class CustomerComponent implements OnInit, OnDestroy {
       {
         entity: null,
         customer,
-        title: "FORMULAIRE D'AJOUT DE TIERS PAYANT ",
+        title: "FORMULAIRE D'AJOUT DE TIERS PAYANT "
       },
       (resp: ICustomer) => {
         if (resp) {
           this.loadPage();
         }
       },
-      'xl',
+      "xl"
     );
   }
 
@@ -406,14 +403,14 @@ export class CustomerComponent implements OnInit, OnDestroy {
       {
         entity: clientTiersPayant,
         customer,
-        title: 'FORMULAIRE DE MODIFICATION DE TIERS PAYANT [ ' + clientTiersPayant.tiersPayantName + ' ]',
+        title: "FORMULAIRE DE MODIFICATION DE TIERS PAYANT [ " + clientTiersPayant.tiersPayantName + " ]"
       },
       (resp: ICustomer) => {
         if (resp) {
           this.loadPage();
         }
       },
-      'xl',
+      "xl"
     );
   }
 
@@ -424,16 +421,16 @@ export class CustomerComponent implements OnInit, OnDestroy {
   }
 
   protected displayTiersPayantAddBtn(customer: ICustomer): boolean {
-    return customer.typeTiersPayant === 'ASSURANCE' && customer.tiersPayants && customer.tiersPayants.length < 4;
+    return customer.typeTiersPayant === "ASSURANCE" && customer.tiersPayants && customer.tiersPayants.length < 4;
   }
 
   private handleNavigation(): void {
     combineLatest(this.activatedRoute.data, this.activatedRoute.queryParamMap, (data: Data, params: ParamMap) => {
-      const page = params.get('page');
+      const page = params.get("page");
       const pageNumber = page !== null ? +page : 1;
-      const sort = (params.get('sort') ?? data['defaultSort']).split(',');
+      const sort = (params.get("sort") ?? data["defaultSort"]).split(",");
       const predicate = sort[0];
-      const ascending = sort[1] === 'asc';
+      const ascending = sort[1] === "asc";
       if (pageNumber !== this.page || predicate !== this.predicate || ascending !== this.ascending) {
         this.predicate = predicate;
         this.ascending = ascending;
@@ -445,15 +442,15 @@ export class CustomerComponent implements OnInit, OnDestroy {
   }
 
   private onSuccess(data: ICustomer[] | null, headers: HttpHeaders, page: number, navigate: boolean): void {
-    this.totalItems = Number(headers.get('X-Total-Count'));
+    this.totalItems = Number(headers.get("X-Total-Count"));
     this.page = page;
     if (navigate) {
-      this.router.navigate(['/customer'], {
+      this.router.navigate(["/customer"], {
         queryParams: {
           page: this.page,
           size: this.itemsPerPage,
-          sort: this.predicate + ',' + (this.ascending ? 'asc' : 'desc'),
-        },
+          sort: this.predicate + "," + (this.ascending ? "asc" : "desc")
+        }
       });
     }
     this.customers = data || [];
@@ -475,12 +472,10 @@ export class CustomerComponent implements OnInit, OnDestroy {
       },
       error: error => {
         this.spinner().hide();
-        this.messageService.add({
-          severity: 'error',
-          summary: 'Erreur',
-          detail: error.error?.errorKey ? this.translate.instant(error.error.errorKey) : error.error?.title || 'Erreur interne du serveur.',
-        });
-      },
+        this.notificationService.error(
+          error.error?.errorKey ? this.translate.instant(error.error.errorKey) : error.error?.title || "Erreur interne du serveur."
+        );
+      }
     });
   }
 }
